Backend/requirements.txt
appdirs==1.4.4
black==21.5b1
click==7.1.2
config==0.5.0.post0
efficientnet-pytorch==0.7.1
fastapi==0.65.1
h11==0.12.0
mypy-extensions==0.4.3
numpy==1.20.3
pathspec==0.8.1
Pillow==8.2.0
pydantic==1.8.2
python-multipart==0.0.5
regex==2021.4.4
six==1.16.0
starlette==0.14.2
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
uvicorn==0.13.4
